requirements.txt
streamlit
pandas
numpy
matplotlib
supabase
plotly
python-dotenv